+++ src/benchmark_main.cpp
@@ -1,0 +1,11 @@
+#include "moveit_middleware_benchmark/scenarios/scenario_perception_pipeline.hpp"
+
+int main(int argc, char** argv)
+{
+  rclcpp::init(argc, argv);
+  benchmark::Initialize(&argc, argv);
+  benchmark::RunSpecifiedBenchmarks();
+  benchmark::Shutdown();
+  rclcpp::shutdown();
+  return 0;
+}
